Improve tchannel outbound introspection
